CLI Support
The following code changes allows FODLego to be used with the CLI.

In ElementaryClasses.py and globaldata.py some code has been refactored
in order to not depend on the elements2.0 file anymore. The period/row
finding is incomplete at the moment, but can be easily completed if
FODLego ever reaches higher periods in the periodic table.