Atomic FODs setup
- Had to switch the values in class Atom since the mValCount was referring
to the variable mBonds, which had to be declared beforehand.

- Also, now there won't be an error for BondOrders if there are not
bonds at all.

- In GlobalData, there must be some convention made for each set of
up/down FODs. Currently we are limited to 'point' and 'tetra', but we
could do something like 'sp3sp2', to refer to up and down, and each of
these could have a transformation for the each shell.