experiment-continuation
This finds some 'optimal' position for the shells of the same principal
quantum number. It is similar to the optimization done in FODMc, but the
idea here is that once it's done, you don't do it again, since the
positions are assumed to have a predetermined optimal position (almost
like a lego).